[refactor] responseCode 처리 누락된 부분 보충
