+++ src/main/java/com/amp/domain/notification/dto/request/FcmTopicMessageRequest.java
@@ -1,4 +1,4 @@
-package com.amp.domain.notification.dto;
+package com.amp.domain.notification.dto.request;
 
 public record FcmTopicMessageRequest(
      String topic,
