+++ src/main/java/com/amp/domain/notification/dto/request/FcmTopicSubscribeRequest.java
@@ -1,4 +1,4 @@
-package com.amp.domain.notification.dto;
+package com.amp.domain.notification.dto.request;
 
 public record FcmTopicSubscribeRequest(
         String fcmToken
